コミットし忘れ@城戸

+++ AnimalForest/Assets/Scripts/Gimmick/GimmickManager.cs
@@ -9,7 +9,7 @@ public class GimmickManager : MonoBehaviour
     [System.NonSerialized]
     public int cost;
     [System.NonSerialized]
-    public float power;
+    public int power;
     [System.NonSerialized]
     public float gimmicktime;
 }
